+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Removals
 Additions
 - Added AutoInlineCenter and AutoInlineBottom anchors
 - Added UiAnimation system
+- Added AddCustomStyle and ApplyCustomStyle to UiStyle to allow for easy custom styling of elements
 
 Improvements
 - Increased Element area calculation recursion limit to 64